// src/types.ts
export interface ExperienceItem {
  title: string;
  company: string;
  period: string;
  description: string[];
}

export interface ProjectItem {
  title: string;
  category: string;
  image: string;
  url: string;
}

export interface ProfileData {
  name: string;
  title: string;
  tagline: string;
  email: string;
  location: string;
  linkedin: string;
  github: string;
  twitter: string;
  instagram: string;
  youtube: string;
  profileImage: string;
}